requirements.txt
annotated-types==0.6.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Cors==4.0.0
flask-openapi3==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
logger==1.4
MarkupSafe==2.1.3
nose2==0.14.0
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
pandasql==0.7.3
pydantic==2.4.2
pydantic_core==2.10.1
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
